requirements.txt
schedule>=1.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
tweepy>=4.14.0
